requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-dotenv==0.16.0
requests==2.25.1
scaleapi==1.0.4
urllib3==1.26.4
